refactor(sui-genesis-builder): split stardust modules
