+++ crates/sui-genesis-builder/src/stardust/mod.rs
@@ -1,0 +1,6 @@
+//! The [`stardust`] module incorporates all the logic necessary for
+//! parsing Stardust UTXOs from a full-snapshot file, and converting
+//! them to the appropriate genesis objects.
+pub mod error;
+pub mod parse;
+pub mod types;
